+++ mcp-test/src/main/java/io/modelcontextprotocol/util/ToolsUtils.java
@@ -1,4 +1,4 @@
-package io.modelcontextprotocol.utils;
+package io.modelcontextprotocol.util;
 
 import io.modelcontextprotocol.spec.McpSchema;
